Limit estimatefee using daemon's relayfee

+++ src/query.rs
@@ -503,7 +503,7 @@ impl Query {
     }
 
     // Fee rate [BTC/kB] to be confirmed in `blocks` from now.
-    pub fn estimate_fee(&self, blocks: usize) -> f32 {
+    pub fn estimate_fee(&self, blocks: usize) -> f64 {
         let mut total_vsize = 0u32;
         let mut last_fee_rate = 0.0;
         let blocks_in_vbytes = (blocks * 1_000_000) as u32; // assume ~1MB blocks
@@ -514,7 +514,7 @@ impl Query {
                 break; // under-estimate the fee rate a bit
             }
         }
-        last_fee_rate * 1e-5 // [BTC/kB] = 10^5 [sat/B]
+        (last_fee_rate as f64) * 1e-5 // [BTC/kB] = 10^5 [sat/B]
     }
 
     pub fn get_banner(&self) -> Result<String> {
